+++ src/classes_body_vehicles.py
@@ -99,7 +99,7 @@ class Vehicle(Base_animated_object):
     def is_collision(self, list_with_units, coord):
     # return True if collision with other object occurs
         for unit in list_with_units:
-            if unit.unit_type == "land":
+            if unit.unit_type == "land" or unit.unit_type == "navy":
                 dist = dist_two_points(coord, unit.coord)
                 if dist < self.hit_box_radius + unit.hit_box_radius and dist > 5: # dist > 5 is to avoid a collision with yourself
                     return True